Refactor: async functions in one file/module
