+++ frontend/src/pages/home/tabs/PostFeed.tsx
@@ -1,7 +1,7 @@
 import Loading from "../../loading/Loading";
 import styles from "./PostFeed.module.css";
 import Button from "../../../components/Button";
-import fetchPosts from "./fetchPostsFunction";
+import { fetchPosts } from "./asyncFunctions";
 import { useQuery } from "@tanstack/react-query";
 
 type Post = {
